Reorganized structure.
