+++ crates/cyancia_shader_graph/src/editor/slot.rs
@@ -1,0 +1,252 @@
+use std::{any::Any, collections::HashMap};
+
+use cyancia_id::Id;
+use iced_core::{
+    Border, Color, Element, Length, Point, Rectangle, Size, Widget,
+    alignment::Vertical,
+    layout, mouse,
+    renderer::{self, Quad},
+    widget::{Operation, Tree, tree},
+};
+use iced_widget::{row, text};
+
+use crate::{
+    GraphRenderer, GraphTheme,
+    graph::slot::{ErasedGraphLiteralUpdateMessage, GraphInputSlotData, GraphOutputSlotData},
+};
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
+pub enum GraphSlotId {
+    Input(Id<GraphInputSlotData>),
+    Output(Id<GraphOutputSlotData>),
+}
+
+impl From<Id<GraphInputSlotData>> for GraphSlotId {
+    fn from(value: Id<GraphInputSlotData>) -> Self {
+        GraphSlotId::Input(value)
+    }
+}
+
+impl From<Id<GraphOutputSlotData>> for GraphSlotId {
+    fn from(value: Id<GraphOutputSlotData>) -> Self {
+        GraphSlotId::Output(value)
+    }
+}
+
+#[derive(Default)]
+pub struct GraphSlotPinPositionCollection {
+    slots: HashMap<GraphSlotId, Point>,
+}
+
+impl Operation for GraphSlotPinPositionCollection {
+    fn traverse(&mut self, operate: &mut dyn FnMut(&mut dyn Operation<()>)) {
+        operate(self);
+    }
+
+    fn custom(&mut self, id: Option<&iced_widget::Id>, bounds: Rectangle, state: &mut dyn Any) {
+        if let Some(state) = state.downcast_ref::<SlotPinState>() {
+            self.slots.insert(state.id.clone(), bounds.center());
+        }
+    }
+}
+
+impl GraphSlotPinPositionCollection {
+    pub fn get(&self, slot_id: &GraphSlotId) -> Option<&Point> {
+        self.slots.get(slot_id)
+    }
+
+    pub fn get_input(&self, slot_id: &Id<GraphInputSlotData>) -> Option<&Point> {
+        self.slots.get(&GraphSlotId::Input(*slot_id))
+    }
+
+    pub fn get_output(&self, slot_id: &Id<GraphOutputSlotData>) -> Option<&Point> {
+        self.slots.get(&GraphSlotId::Output(*slot_id))
+    }
+
+    pub fn clear(&mut self) {
+        self.slots.clear();
+    }
+
+    pub fn all(&self) -> impl Iterator<Item = (&GraphSlotId, &Point)> {
+        self.slots.iter()
+    }
+}
+
+pub struct SlotPin {
+    pub id: GraphSlotId,
+    pub radius: f32,
+    pub color: Color,
+}
+
+impl<Message, Theme, Renderer> Widget<Message, Theme, Renderer> for SlotPin
+where
+    Renderer: iced_core::Renderer,
+{
+    fn size(&self) -> Size<Length> {
+        let d = self.radius * 2.0;
+        Size::new(Length::Fixed(d), Length::Fixed(d))
+    }
+
+    fn layout(
+        &mut self,
+        tree: &mut Tree,
+        renderer: &Renderer,
+        limits: &layout::Limits,
+    ) -> layout::Node {
+        let d = self.radius * 2.0;
+        layout::atomic(limits, d, d)
+    }
+
+    fn draw(
+        &self,
+        tree: &Tree,
+        renderer: &mut Renderer,
+        theme: &Theme,
+        style: &renderer::Style,
+        layout: layout::Layout<'_>,
+        cursor: mouse::Cursor,
+        viewport: &Rectangle,
+    ) {
+        renderer.fill_quad(
+            Quad {
+                bounds: layout.bounds(),
+                border: Border::default().rounded(self.radius),
+                ..Default::default()
+            },
+            self.color,
+        );
+    }
+
+    fn operate(
+        &mut self,
+        tree: &mut Tree,
+        layout: layout::Layout<'_>,
+        renderer: &Renderer,
+        operation: &mut dyn Operation,
+    ) {
+        operation.custom(None, layout.bounds(), &mut SlotPinState { id: self.id });
+    }
+}
+
+struct SlotPinState {
+    id: GraphSlotId,
+}
+pub enum SlotSide {
+    Left,
+    Right,
+}
+
+pub fn input_slot(
+    slot_id: Id<GraphInputSlotData>,
+    slot: &GraphInputSlotData,
+) -> Element<'static, ErasedGraphLiteralUpdateMessage, GraphTheme, GraphRenderer> {
+    match slot.connected {
+        Some(_) => empty_slot(
+            slot_id.into(),
+            slot.data.ty().color(),
+            slot.name,
+            SlotSide::Left,
+        ),
+        None => valued_slot(
+            slot_id.into(),
+            slot.data.ty().color(),
+            slot.name,
+            SlotSide::Left,
+            slot.data.ty().view_literal(slot_id, slot.data.value()),
+        ),
+    }
+}
+
+pub fn output_slot<'a, Message>(
+    slot_id: Id<GraphOutputSlotData>,
+    slot: &GraphOutputSlotData,
+) -> Element<'a, Message, GraphTheme, GraphRenderer>
+where
+    Message: 'a,
+{
+    empty_slot(
+        slot_id.into(),
+        slot.data.ty().color(),
+        slot.name,
+        SlotSide::Right,
+    )
+}
+
+pub fn empty_slot<'a, Message, Theme, Renderer>(
+    id: GraphSlotId,
+    color: Color,
+    name: &'a str,
+    slot_side: SlotSide,
+) -> Element<'a, Message, Theme, Renderer>
+where
+    Message: 'a,
+    Theme: text::Catalog + 'a,
+    Renderer: iced_core::renderer::Renderer + iced_core::text::Renderer + 'a,
+{
+    let text = text(name);
+    let pin = Element::new(SlotPin {
+        id,
+        color,
+        radius: 3.0,
+    });
+
+    match slot_side {
+        SlotSide::Left => row![pin, text].align_y(Vertical::Center).spacing(4).into(),
+        SlotSide::Right => row![text, pin].align_y(Vertical::Center).spacing(4).into(),
+    }
+}
+
+pub fn valued_slot_unconnectable<'a, Message, Theme, Renderer>(
+    name: &'a str,
+    slot_side: SlotSide,
+    widget: Element<'a, Message, Theme, Renderer>,
+) -> Element<'a, Message, Theme, Renderer>
+where
+    Message: 'a,
+    Theme: text::Catalog + 'a,
+    Renderer: iced_core::renderer::Renderer + iced_core::text::Renderer + 'a,
+{
+    let text = text(name);
+
+    match slot_side {
+        SlotSide::Left => row![text, widget]
+            .align_y(Vertical::Center)
+            .spacing(4)
+            .into(),
+        SlotSide::Right => row![widget, text]
+            .align_y(Vertical::Center)
+            .spacing(4)
+            .into(),
+    }
+}
+
+pub fn valued_slot<'a, Message, Theme, Renderer>(
+    id: GraphSlotId,
+    color: Color,
+    name: &'a str,
+    slot_side: SlotSide,
+    widget: Element<'a, Message, Theme, Renderer>,
+) -> Element<'a, Message, Theme, Renderer>
+where
+    Message: 'a,
+    Theme: text::Catalog + 'a,
+    Renderer: iced_core::renderer::Renderer + iced_core::text::Renderer + 'a,
+{
+    let text = text(name);
+    let pin = Element::new(SlotPin {
+        id,
+        color,
+        radius: 3.0,
+    });
+
+    match slot_side {
+        SlotSide::Left => row![pin, text, widget]
+            .align_y(Vertical::Center)
+            .spacing(4)
+            .into(),
+        SlotSide::Right => row![widget, text, pin]
+            .align_y(Vertical::Center)
+            .spacing(4)
+            .into(),
+    }
+}
